% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create a reduced segment file (integreates other functions)}
\usage{
create()
}
\arguments{
\item{none}{no arguments yet}
}
\value{
Nothing returned yet
}
\description{
Create a reduced segment file (integreates other functions)
}
\examples{
create()
}
