% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNSeg.R
\name{reduceSegment}
\alias{reduceSegment}
\title{Core: Creates a filtered consensus segments file based on a traditional segments file}
\usage{
reduceSegment(sourceData, namesArray, filterThreshold)
}
\arguments{
\item{sourceData}{(dataframe or matrix) the original segments table}

\item{namesArray}{(array) contains 5 elements, in order:
namesArray[1]: The column name identifying the segment's chromosome
namesArray[2]: The column name identifying the genomic start location
namesArray[3]: The column name identifying the genomic stop location
namesArray[4]: The column name identifying the mean log2 ratio of the segment
namesArray[5]: The column name identifying the sample that the segment represents}

\item{filterThreshold}{(float) between 0 and 1.0, the threshold for filtering based on mean absolute deviation}
}
\value{
A filtered dataframe of consensus segments
}
\description{
Core: Creates a filtered consensus segments file based on a traditional segments file
}
\examples{
# Read the raw segment data
data(segData)
segData <- subset(segData, segData$Type == "TotalCN")
reduceSegment(segData, 
c("Chromosome", "StartPosition", "StopPosition", 
"Median.Log2.Ratio", "FileName"))
}
