% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_segments.R
\name{import_segments}
\alias{import_segments}
\title{Imports the inital variables from a segments file}
\usage{
import_segments()
}
\arguments{
\item{interactive_input_path}{A full file path to the tab-delineated imput file}

\item{more}{Will add more}
}
\value{
chromosome The "chromosome" variable is created
}
\description{
Imports the inital variables from a segments file
}
\examples{
import_segments()
}
